v0.4.0 (#33)
**Change Log:**
* Increase Prompt API timeout to 12 seconds
* Fixed validator whitelisting
* Updated subnet protocol
* Added Fetch API for miners (part of the prompting API)
* Added remote logging of the analysis metrics (to build dashboards for miners and measure the overall subnet performance)
* Deprecated YARA engine for prompt injection analyzer
* Enhanced message signing between validators and miners
* Deprecated local fallback for the creation of prompts
* Added "Sensitive Information" analyzer
* Added "Token Classification" engine to Sensitive Information Analyzer
* Deprecated documentation in GitHub repository in favor of separate documentation repository (https://docs.synapsec.ai) 
* Upgraded minimum Bittensor version to 6.9.3
* Added vector search engine result validation
* Modified base penalties
