+++ .gitignore
@@ -157,7 +157,7 @@ cython_debug/
 #  be found at https://github.com/github/gitignore/blob/main/Global/JetBrains.gitignore
 #  and can be added to the global gitignore or merged into this file.  For a more nuclear
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
-#.idea/
+.idea/
 
 .vscode
 
@@ -168,3 +168,6 @@ llm-defender*.config.js
 
 # Wandb
 wandb
+
+# chromadb
+chroma
