+++ docker-compose.yml
@@ -1,6 +1,6 @@
 version: '3'
 services:
-  validator:
+  validator1:
     image: llm-defender-subnet
     build:
       context: .
@@ -11,6 +11,18 @@ services:
     networks:
       llm-defender-subnet-network-development:
         ipv4_address: ${VALIDATOR_IP}
+  
+  # validator2:
+  #   image: llm-defender-subnet
+  #   build:
+  #     context: .
+  #     dockerfile: Dockerfile   
+  #   command: /bin/bash -c "source /tmp/.venv/bin/activate && pip install -e ${VALIDATOR_PIP_INSTALL_PATH} && python3 /var/run/llm-defender-subnet/llm_defender/neurons/validator.py --netuid ${NETUID} --subtensor.network ${SUBTENSOR_NETWORK} --wallet.path ${WALLET_PATH} --wallet.name ${VALIDATOR2_WALLET} --wallet.hotkey ${VALIDATOR2_HOTKEY} --logging.debug"
+  #   volumes:
+  #     - .:/var/run/llm-defender-subnet
+  #   networks:
+  #     llm-defender-subnet-network-development:
+  #       ipv4_address: ${VALIDATOR2_IP}
 
   miner1:
     image: llm-defender-subnet
@@ -19,7 +31,7 @@ services:
       dockerfile: Dockerfile
     volumes:
       - .:/var/run/llm-defender-subnet
-    command: /bin/bash -c "source /tmp/.venv/bin/activate && pip install -e ${MINER_PIP_INSTALL_PATH} && pip uninstall -y uvloop && python3 /var/run/llm-defender-subnet/llm_defender/neurons/miner.py --netuid ${NETUID} --subtensor.network ${SUBTENSOR_NETWORK} --wallet.path ${WALLET_PATH} --wallet.name ${MINER_WALLET} --wallet.hotkey ${MINER_HOTKEY} --logging.debug --axon.port ${MINER_PORT} --validator_min_stake 0 --axon.external_ip ${MINER_IP}"
+    command: /bin/bash -c "source /tmp/.venv/bin/activate && pip install -e ${MINER_PIP_INSTALL_PATH} && pip uninstall -y uvloop && python3 /var/run/llm-defender-subnet/llm_defender/neurons/miner.py --netuid ${NETUID} --subtensor.network ${SUBTENSOR_NETWORK} --wallet.path ${WALLET_PATH} --wallet.name ${MINER_WALLET} --wallet.hotkey ${MINER_HOTKEY} --axon.port ${MINER_PORT} --validator_min_stake 0 --axon.external_ip ${MINER_IP}"
     networks:
       llm-defender-subnet-network-development:
         ipv4_address: ${MINER_IP}
@@ -30,7 +42,7 @@ services:
       dockerfile: Dockerfile
     volumes:
       - .:/var/run/llm-defender-subnet
-    command: /bin/bash -c "source /tmp/.venv/bin/activate && pip install -e ${MINER_PIP_INSTALL_PATH} && pip uninstall -y uvloop && python3 /var/run/llm-defender-subnet/llm_defender/neurons/miner.py --netuid ${NETUID} --subtensor.network ${SUBTENSOR_NETWORK} --wallet.path ${WALLET_PATH} --wallet.name ${MINER2_WALLET} --wallet.hotkey ${MINER2_HOTKEY} --logging.debug --axon.port ${MINER2_PORT} --validator_min_stake 0 --axon.external_ip ${MINER2_IP}"
+    command: /bin/bash -c "source /tmp/.venv/bin/activate && pip install -e ${MINER_PIP_INSTALL_PATH} && pip uninstall -y uvloop && python3 /var/run/llm-defender-subnet/llm_defender/neurons/miner.py --netuid ${NETUID} --subtensor.network ${SUBTENSOR_NETWORK} --wallet.path ${WALLET_PATH} --wallet.name ${MINER2_WALLET} --wallet.hotkey ${MINER2_HOTKEY} --axon.port ${MINER2_PORT} --validator_min_stake 0 --axon.external_ip ${MINER2_IP}"
     networks:
       llm-defender-subnet-network-development:
         ipv4_address: ${MINER2_IP}
