+++ llm_defender/base/utils.py
@@ -324,7 +324,7 @@ def validate_signature(hotkey: str, data: str, signature: str) -> bool:
         bt.logging.error(f'Failed to validate signature: {signature} for data: {data} with error: {e}')
         return False
 
-def sign_data(wallet: bt.wallet, data: str) -> str:
+def sign_data(hotkey: bt.Keypair, data: str) -> str:
     """Signs the given data with the wallet hotkey
     
     Arguments:
@@ -338,13 +338,13 @@ def sign_data(wallet: bt.wallet, data: str) -> str:
             Signature of the key signing for the data
     """
     try:
-        signature = wallet.hotkey.sign(data.encode()).hex()
+        signature = hotkey.sign(data.encode()).hex()
         return signature
     except TypeError as e:
-        bt.logging.error(f'Unable to sign data: {data} with wallet hotkey: {wallet.hotkey} due to error: {e}')
+        bt.logging.error(f'Unable to sign data: {data} with wallet hotkey: {hotkey.ss58_address} due to error: {e}')
         raise TypeError from e
     except AttributeError as e:
-        bt.logging.error(f'Unable to sign data: {data} with wallet hotkey: {wallet.hotkey} due to error: {e}')
+        bt.logging.error(f'Unable to sign data: {data} with wallet hotkey: {hotkey.ss58_address} due to error: {e}')
         raise AttributeError from e
 
 def validate_prompt(prompt_dict):
@@ -353,7 +353,6 @@ def validate_prompt(prompt_dict):
     key_types = {
     'analyzer':str,
     'category':str,
-    'prompt':str,
     'label':int,
     'weight':(int, float),
     'hotkey': str,
@@ -366,7 +365,7 @@ def validate_prompt(prompt_dict):
     if len([pd for pd in prompt_dict]) != 8:
         return False
     for pd in prompt_dict:
-        if pd not in ['analyzer','category','prompt','label','weight', 'created_at', 'synapse_uuid', 'hotkey']:
+        if pd not in ['analyzer','category','label','weight', 'created_at', 'synapse_uuid', 'hotkey']:
             return False 
         if not isinstance(prompt_dict[pd], key_types[pd]):
             return False
