+++ llm_defender/core/miners/analyzers/prompt_injection/analyzer.py
@@ -3,35 +3,27 @@ import time
 import secrets
 import bittensor as bt
 from llm_defender.base.protocol import LLMDefenderProtocol
-from llm_defender.core.miners.analyzers.prompt_injection.yara import YaraEngine
 from llm_defender.core.miners.analyzers.prompt_injection.text_classification import TextClassificationEngine
 from llm_defender.core.miners.analyzers.prompt_injection.vector_search import VectorEngine
 from llm_defender.base.utils import sign_data
 
 # Load wandb library only if it is enabled
 from llm_defender import __wandb__ as wandb
-if wandb is True:
-    from llm_defender.base.wandb_handler import WandbHandler
 
 class PromptInjectionAnalyzer:
     """This class is responsible for handling the analysis for prompt injection
     
-    The PromptInjectionAnalyzer class contains all of the code for a Miner neuron
+    The PromptInjectionAnalyzer class contains all the code for a Miner neuron
     to generate a confidence score for a Prompt Injection Attack.
 
     Attributes:
         chromadb_client:
-            Stores the 'clint' output from VectorEngine.initialize() This is from:
+            Stores the 'client' output from VectorEngine.initialize() This is from:
             llm_defender/core/miners/engines/prompt_injection/vector_search.py
         model:
             Stores the 'model' output for an engine.
         tokenizer:
             Stores the 'tokenized' output for an engine.
-        yara_rules:
-            Stores the 'rules' output of YaraEngine.initialize() This is only when using
-            the YaraEngine, located at:
-
-            llm_defender/core/miners/engines/prompt_injection/yara.py
 
     Methods:
         execute:
@@ -49,7 +41,6 @@ class PromptInjectionAnalyzer:
         # Configuration options for the analyzer
         self.chromadb_client = VectorEngine().initialize()
         self.model, self.tokenizer = TextClassificationEngine().initialize()
-        self.yara_rules = YaraEngine().initialize()
 
         # Enable wandb if it has been configured
         if wandb is True:
@@ -58,30 +49,24 @@ class PromptInjectionAnalyzer:
         else:
             self.wandb_enabled = False
             self.wandb_handler = None
-
     
-    def execute(self, synapse: LLMDefenderProtocol) -> dict:
+    def execute(self, synapse: LLMDefenderProtocol, prompt: str) -> dict:
         # Responses are stored in a dict
-        output = {"analyzer": "Prompt Injection", "prompt": synapse.prompt, "confidence": None, "engines": []}
+        output = {"analyzer": "Prompt Injection", "confidence": None, "engines": []}
+
+        prompt = prompt
 
         engine_confidences = []
 
-        # Execute YARA engine
-        yara_engine = YaraEngine(prompt=synapse.prompt)
-        yara_engine.execute(rules=self.yara_rules)
-        yara_response = yara_engine.get_response().get_dict()
-        output["engines"].append(yara_response)
-        engine_confidences.append(yara_response["confidence"])
-
         # Execute Text Classification engine
-        text_classification_engine = TextClassificationEngine(prompt=synapse.prompt)
+        text_classification_engine = TextClassificationEngine(prompt=prompt)
         text_classification_engine.execute(model=self.model, tokenizer=self.tokenizer)
         text_classification_response = text_classification_engine.get_response().get_dict()
         output["engines"].append(text_classification_response)
         engine_confidences.append(text_classification_response["confidence"])
 
         # Execute Vector Search engine
-        vector_engine = VectorEngine(prompt=synapse.prompt)
+        vector_engine = VectorEngine(prompt=prompt)
         vector_engine.execute(client=self.chromadb_client)
         vector_response = vector_engine.get_response().get_dict()
         output["engines"].append(vector_response)
@@ -95,28 +80,28 @@ class PromptInjectionAnalyzer:
         output["synapse_uuid"] = synapse.synapse_uuid
         output["nonce"] = secrets.token_hex(24)
         output["timestamp"] = str(int(time.time()))
-        
-        data_to_sign = f'{output["synapse_uuid"]}{output["nonce"]}{output["timestamp"]}'
+
+        data_to_sign = f'{output["synapse_uuid"]}{output["nonce"]}{self.wallet.hotkey.ss58_address}{output["timestamp"]}'
 
         # Generate signature for the response
-        output["signature"] = sign_data(self.wallet, data_to_sign)
+        output["signature"] = sign_data(self.wallet.hotkey, data_to_sign)
 
         # Wandb logging
         if self.wandb_enabled:
             self.wandb_handler.set_timestamp()
 
             wandb_logs = [
-                {f"{self.miner_uid}:{self.miner_hotkey}_YARA Confidence":yara_response['confidence']},
-                {f"{self.miner_uid}:{self.miner_hotkey}_Text Classification Confidence":text_classification_response['confidence']},
-                {f"{self.miner_uid}:{self.miner_hotkey}_Vector Search Confidence":vector_response['confidence']},
-                {f"{self.miner_uid}:{self.miner_hotkey}_Total Confidence":output['confidence']}
-            ]   
+                {f"{self.miner_uid}:{self.miner_hotkey}_Text Classification Confidence": text_classification_response['confidence']},
+                {f"{self.miner_uid}:{self.miner_hotkey}_Vector Search Confidence": vector_response['confidence']},
+                {f"{self.miner_uid}:{self.miner_hotkey}_Total Confidence": output['confidence']}
+            ]
 
             for wandb_log in wandb_logs:
                 self.wandb_handler.log(data=wandb_log)
-            
+
             bt.logging.trace(f"Wandb logs added: {wandb_logs}")
-        
+
+        bt.logging.debug(f'Setting synapse.output to: {output}')
         synapse.output = output
 
         return output
