+++ llm_defender/core/miners/miner.py
@@ -14,15 +14,22 @@ from typing import Tuple
 import sys
 import requests
 import bittensor as bt
+import secrets
+import time
 from llm_defender.base.neuron import BaseNeuron
 from llm_defender.base.protocol import LLMDefenderProtocol
-from llm_defender.base.utils import validate_miner_blacklist, validate_signature
+from llm_defender.core.miners.analyzers import SupportedAnalyzers
+from llm_defender.base.utils import validate_miner_blacklist, validate_signature, sign_data
 from llm_defender.core.miners.analyzers.prompt_injection.analyzer import (
     PromptInjectionAnalyzer,
 )
 
 # Load wandb library only if it is enabled
 from llm_defender import __wandb__ as wandb
+from llm_defender.core.miners.analyzers.sensitive_information.analyzer import (
+    SensitiveInformationAnalyzer
+)
+
 if wandb is True:
     from llm_defender.base.wandb_handler import WandbHandler
 
@@ -112,11 +119,14 @@ class LLMDefenderMiner(BaseNeuron):
         else:
             self.wandb_enabled = False
             self.wandb_handler = None
-        
+
         # Initialize the analyzers
         self.analyzers = {
-            "Prompt Injection": PromptInjectionAnalyzer(
-                wallet=self.wallet, subnet_version=self.subnet_version, wandb_handler=self.wandb_handler,miner_uid = self.miner_uid
+            str(SupportedAnalyzers.PROMPT_INJECTION): PromptInjectionAnalyzer(
+                wallet=self.wallet, subnet_version=self.subnet_version, wandb_handler=self.wandb_handler, miner_uid=self.miner_uid
+            ),
+            str(SupportedAnalyzers.SENSITIVE_INFORMATION): SensitiveInformationAnalyzer(
+                wallet=self.wallet, subnet_version=self.subnet_version, wandb_handler=self.wandb_handler, miner_uid=self.miner_uid
             )
         }
 
@@ -175,6 +185,25 @@ class LLMDefenderMiner(BaseNeuron):
         bt.logging.info(f"Miner is running with UID: {miner_uid}")
 
         return wallet, subtensor, metagraph, miner_uid
+    
+    def get_prompt_from_api(self, hotkey, signature, synapse_uuid, timestamp, nonce) -> dict:
+        """Retrieves a prompt from the prompt API"""
+
+        headers = {
+            "X-Hotkey": hotkey,
+            "X-Signature": signature,
+            "X-SynapseUUID": synapse_uuid,
+            "X-Timestamp": timestamp,
+            "X-Nonce": nonce,
+            "X-Version": str(self.subnet_version),
+            "X-API-Key": hotkey
+        }
+
+        res = self.requests_post(url="https://fetch-api.synapsec.ai/fetch", headers=headers, data={}, timeout=12)
+        
+        if res and "prompt" in res.keys():
+            return res["prompt"]
+        return None
 
     def check_whitelist(self, hotkey):
         """
@@ -321,8 +350,13 @@ class LLMDefenderMiner(BaseNeuron):
                 f"Received a synapse from a validator with higher subnet version ({synapse.subnet_version}) than yours ({self.subnet_version}). Please update the miner."
             )
 
+        # Validate that nonce has not been reused
+        if not self.validate_nonce(synapse.synapse_nonce):
+            bt.logging.warning(f'Received a synapse with previous used nonce: {synapse}')
+            return synapse
+        
         # Synapse signature verification
-        data = f'{synapse.synapse_uuid}{synapse.synapse_nonce}{synapse.synapse_timestamp}'
+        data = f'{synapse.synapse_uuid}{synapse.synapse_nonce}{synapse.dendrite.hotkey}{synapse.synapse_timestamp}'
         if not validate_signature(
             hotkey=synapse.dendrite.hotkey,
             data=data,
@@ -337,25 +371,44 @@ class LLMDefenderMiner(BaseNeuron):
                 f"Succesfully validated signature for the synapse. Hotkey: {synapse.dendrite.hotkey}, data: {data}, signature: {synapse.synapse_signature}"
             )
 
+        # Get the prompt to be analyzed
+        nonce = str(secrets.token_hex(24))
+        timestamp = str(int(time.time()))
+
+        data = f'{synapse.synapse_uuid}{nonce}{timestamp}'
+
+        prompt = self.get_prompt_from_api(
+            hotkey=self.wallet.hotkey.ss58_address,
+            signature=sign_data(hotkey=self.wallet.hotkey, data=data),
+            synapse_uuid=synapse.synapse_uuid, 
+            timestamp=timestamp, 
+            nonce=nonce
+        )
+
         # Execute the correct analyzer
-        if synapse.analyzer == "Prompt Injection":
-            bt.logging.debug(f"Executing the {synapse.analyzer} analyzer")
-            output = self.analyzers["Prompt Injection"].execute(synapse=synapse)
-        else:
+        if not SupportedAnalyzers.is_valid(synapse.analyzer):
             bt.logging.error(
                 f"Unable to process synapse: {synapse} due to invalid analyzer: {synapse.analyzer}"
             )
             return synapse
 
+        bt.logging.debug(f"Executing the {synapse.analyzer} analyzer")
+        output = self.analyzers[synapse.analyzer].execute(synapse=synapse, prompt=prompt)
+
         bt.logging.debug(
-            f'Processed prompt: {output["prompt"]} with analyzer: {output["analyzer"]}'
+            f'Processed prompt "{prompt}" with analyzer: {output["analyzer"]}'
         )
         bt.logging.debug(
             f'Engine data for {output["analyzer"]} analyzer: {output["engines"]}'
         )
-        bt.logging.success(
-            f'Processed synapse from UID: {self.metagraph.hotkeys.index(synapse.dendrite.hotkey)} - Confidence: {output["confidence"]} - UUID: {output["synapse_uuid"]}'
-        )
+        if self.check_whitelist(hotkey=synapse.dendrite.hotkey):
+            bt.logging.success(
+                f'Processed synapse from whitelisted hotkey: {synapse.dendrite.hotkey} - Confidence: {output["confidence"]} - UUID: {output["synapse_uuid"]}'
+            )
+        else:
+            bt.logging.success(
+                f'Processed synapse from UID: {self.metagraph.hotkeys.index(synapse.dendrite.hotkey)} - Confidence: {output["confidence"]} - UUID: {output["synapse_uuid"]}'
+            )
 
         return synapse
