--- llm_defender/core/validators/penalty/__init__.py
@@ -1,1 +1,0 @@
-from . import base, similarity, duplicate
