+++ llm_defender/neurons/miner.py
@@ -98,6 +98,13 @@ def main(miner: LLMDefenderMiner):
 
                     miner.metagraph.sync(subtensor=miner.subtensor)
 
+                    # Check registration status
+                    if miner.wallet.hotkey.ss58_address not in miner.metagraph.hotkeys:
+                        bt.logging.error(f"Hotkey is not registered on metagraph: {miner.wallet.hotkey.ss58_address}.")
+                    
+                    # Save used nonces
+                    miner.save_used_nonces()
+
                 miner.metagraph = miner.subtensor.metagraph(miner.neuron_config.netuid)
                 log = (
                     f"Version:{version} | "
