+++ scripts/fine_tuning_helpers/yara_helper.py
@@ -1,4 +1,4 @@
-from llm_defender.core.miners.analyzers.prompt_injection.yara import YaraEngine
+from llm_defender.core.miners.analyzers.sensitive_information.yara import YaraEngine
 
 engine = YaraEngine()
 engine.prepare()
