+++ scripts/prep.py
@@ -2,19 +2,28 @@
 This script prepares the engines before miner is executed.
 """
 import sys
-from llm_defender.core.miners.analyzers.prompt_injection.yara import YaraEngine
-from llm_defender.core.miners.analyzers.prompt_injection.text_classification import TextClassificationEngine
+
+from llm_defender.core.miners.analyzers.prompt_injection.text_classification import (
+    TextClassificationEngine as PromptTextClassEngine
+)
+from llm_defender.core.miners.analyzers.sensitive_information.yara import YaraEngine
+from llm_defender.core.miners.analyzers.sensitive_information.token_classification import (
+    TokenClassificationEngine as SensitiveTokenClassEngine
+)
 from llm_defender.core.miners.analyzers.prompt_injection.vector_search import VectorEngine
+
 
 def prepare_engines():
     """Prepare the engines"""
-    # Prepare text classification engine
-
-    if not TextClassificationEngine().prepare():
-        print("Unable to prepare text classification engine")
+    # Prepare text classification engines
+    if not PromptTextClassEngine().prepare():
+        print("Unable to prepare text classification engine for prompt injection")
         sys.exit(1)
 
-    print("Prepared Text Classification engine")
+    if not SensitiveTokenClassEngine().prepare():
+        print("Unable to prepare text classification engine for sensitive information")
+        sys.exit(1)
+    print("Prepared Text Classification engines")
 
     # Prepare vector search engine
     if not VectorEngine().prepare():
@@ -23,12 +32,12 @@ def prepare_engines():
 
     print("Prepared Vector Search engine")
 
-    # Prepare YARA engine
-    if not YaraEngine().prepare():
-        print("Unable to prepare vector search engine")
-        sys.exit(1)
+    # # Prepare YARA engine
+    # if not YaraEngine().prepare():
+    #     print("Unable to prepare vector search engine")
+    #     sys.exit(1)
 
-    print("Prepared YARA engine")
+    # print("Prepared YARA engine")
 
 if __name__ == "__main__":
     prepare_engines()
