+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = llm_defender
-version = 0.3.0
+version = 0.4.0
 author = ceterum1
 author_email = ceterum1@proton.me
 description = This project implements the llm-defender subnet for Bittensor
@@ -26,7 +26,7 @@ license_files = LICENSE
 packages = find:
 python_requires = >=3.10
 install_requires =
-    bittensor
+    bittensor >= 6.9.3
     torch
     gitpython
     python-dotenv
@@ -43,3 +43,5 @@ miner =
     yara-python
 validator =
     scikit-learn
+    sentence-transformers
+    numpy
